+++ marginaleffects/hypotheses.py
@@ -7,7 +7,7 @@ from .uncertainty import get_jacobian, get_se, get_z_p_ci
 from .utils import sort_columns
 
 
-def hypotheses(model, hypothesis=None, conf_int=0.95, vcov=True):
+def hypotheses(model, hypothesis=None, conf_level=0.95, vcov=True):
     # sanity checks
     V = sanitize_vcov(vcov, model)
 
@@ -22,6 +22,6 @@ def hypotheses(model, hypothesis=None, conf_int=0.95, vcov=True):
         J = get_jacobian(fun, model.params.to_numpy())
         se = get_se(J, V)
         out = out.with_columns(pl.Series(se).alias("std_error"))
-        out = get_z_p_ci(out, model, conf_int=conf_int)
+        out = get_z_p_ci(out, model, conf_level=conf_level)
     out = sort_columns(out, by=None)
     return out