+++ NEWS.md
@@ -1,5 +1,8 @@
 # Dev
 
+Breaking change:
+
+* Rename argument to match `R` `marginaleffects`: `conf_int` -> `conf_level`
 * `MarginaleffectsDataFrame` class inherits from `pl.DataFrame` for better printing and to host useful attributes.
 
 # 0.0.2